fix #3: compress tips and scripts with DTE
Adding Popslide and DTE compression managed to save only about 1K of
PRG ROM in all.  But decompressing text to fixed-address, fixed-size
buffers actually reduced the pointer juggling in cutscene.s, and now
we have over 2 KiB of stuff moved out to the library bank (plus 1 KiB
for Thwaite's music data if we use an integrated soundtrack).

As for time to make clean && make all, DTE is by far the most
time-consuming part of building Thwaite.